Opening AO from browser or finder will do an install
